+++ tgframework/web/__init__.py
@@ -3,13 +3,11 @@
 """
 
 from .server import WebServer
-from .admin import AdminPanel
 from .auth import TelegramAuth
 from .routing import Router, Controller
 
 __all__ = [
     "WebServer",
-    "AdminPanel",
     "TelegramAuth",
     "Router",
     "Controller",